[Feature: 28] Added support to stop dynamic dependencies at runtime.
Note: The feature is incomplete but usable. For completeness, the list of dynamic dependencies needs to be purged from stale references. The dependencies list should however make it in a way that does not add latency, blocking calls, create a race condition or removes a static dependency from the array stored in the SMO base. Ideally, it would be handled by SMO Base's process to avoid the list of stale dependencies from growing out of control. At the moment, a connection handler type of application would risk running into a gigantic array of dependencies...